requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
python-dotenv>=1.0.0
ipython==9.0.2
docarray>=0.1.0
pypdf>=5.3.1
faiss-cpu>=1.10.0
openai>=1.12.0
tiktoken>=0.5.2
numpy>=1.25.0,<3.0
pandas>=2.0.0
packaging>=24.2
